feat: 커뮤니티 개설 요청 페이지 작성 (tsx -> jsx)
